Initial work on flappy bird

+++ include/minigame/Minigame.h
@@ -18,7 +18,7 @@ namespace Game3 {
 			virtual void tick(UIContext &, double delta) = 0;
 			virtual void reset() = 0;
 			virtual void render(UIContext &, const RendererContext &) = 0;
-			virtual void setSize(float width, float height) = 0;
+			virtual void setSize(int width, int height) = 0;
 
 		protected:
 			std::size_t score{};
